LayoutTests/http/tests/websocket/tests/hybi/non-document-mixed-content-blocked.https-expected.txt

CONSOLE MESSAGE: WebSocket connection failed: The page at https://127.0.0.1:8443/websocket/tests/hybi/non-document-mixed-content-blocked.https.html was blocked from connecting insecurely to ws://127.0.0.1:8880/websocket/tests/hybi/simple either because the protocol is insecure or the page is embedded from an insecure page.
Test mixed content from unsecure WebSocket within Worker blocked from secure page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS worker onerror. readyState 3
DONE
TEST COMPLETE
